+++ adapters/base_adapter.py
@@ -5,9 +5,14 @@ class Adapter():
     def __init__(self, devices):
         self.devices = []
 
+    def convert_message(self, message):
+        return message
+
     def handleMqttMessage(self, device_data, message):
+        converted_message = self.convert_message(message)
+
         for device in self.devices:
-            device.handle_message(device_data, message)
+            device.handle_message(device_data, converted_message)
 
     def handleCommand(self, alias, device, device_data, command, level, color):
         Domoticz.Debug('Update command has not been implemented for device "' + device.Name + '"')